Refactor kernel fns
