+++ src/ops/fused.cpp
@@ -15,7 +15,7 @@
 
 #include "flexflow/ops/fused.h"
 #include "flexflow/model.h"
-#include "flexflow/ops/batch_matmul.h"
+#include "flexflow/ops/kernels/batch_matmul_kernels.h"
 #include "flexflow/ops/batch_norm.h"
 #include "flexflow/ops/dropout.h"
 #include "flexflow/ops/element_unary.h"
@@ -252,7 +252,7 @@ __host__ void FusedOp::forward_task(Task const *task,
           batch *= dim_size;
         }
         BatchMatmulMeta *meta = (BatchMatmulMeta *)metas->meta[op];
-        BatchMatmul::forward_kernel_wrapper(
+        Kernels::BatchMatmul::forward_kernel_wrapper(
             meta,
             my_output_accessor[0].get_float_ptr(),
             my_input_accessor[0].get_float_ptr(),
@@ -638,7 +638,7 @@ __host__ void FusedOp::backward_task(Task const *task,
           batch *= dim_size;
         }
         BatchMatmulMeta *meta = (BatchMatmulMeta *)metas->meta[op];
-        BatchMatmul::backward_kernel_wrapper(
+        Kernels::BatchMatmul::backward_kernel_wrapper(
             meta,
             (float const *)my_output_accessor[0].get_float_ptr(),
             (float const *)my_output_grad_accessor[0].get_float_ptr(),
